# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Account and all owned financial data deleted"
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
...
    "/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "categories"
        ]
      },
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "categories"
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "categories"
        ]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The archived (soft-deleted) category",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "categories"
        ]
      }
    },
    "/transactions": {
      "post": {
        "operationId": "TransactionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "transactions"
        ]
      },
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "income",
                "expense",
                "transfer"
              ]
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z||([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z||([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "transactions"
        ]
      }
    },
    "/transactions/{id}": {
      "get": {
        "operationId": "TransactionsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "transactions"
        ]
      },
      "patch": {
        "operationId": "TransactionsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "transactions"
        ]
      },
      "delete": {
        "operationId": "TransactionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The soft-deleted transaction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "transactions"
        ]
      }
    },
    "/transactions/{id}/restore": {
      "patch": {
        "operationId": "TransactionsController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pocketly API",
    "description": "Core finance domain API for Pocketly: accounts, categories, transactions, budgets, and analysis.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "clerk": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Clerk session token"
      }
    },
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "clerkUserId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "clerkUserId",
          "email",
          "name",
          "currency",
          "timezone",
          "createdAt",
          "updatedAt"
        ]
      },
      "DeleteAccountDto": {
        "type": "object",
        "properties": {
          "confirm": {
            "type": "boolean",
            "const": true,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        },
        "required": [
          "confirm"
        ]
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ]
          },
          "icon": {
            "type": "string",
            "maxLength": 50
          },
          "color": {
            "type": "string",
            "maxLength": 20
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "userId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "name": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "icon": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "color": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "ignored": {
            "type": "boolean",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "deletedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true
          },
          "createdAt": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "updatedAt": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        },
        "required": [
          "_id",
          "userId",
          "name",
          "type",
          "ignored",
          "deletedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ]
          },
          "icon": {
            "type": "string",
            "maxLength": 50
          },
          "color": {
            "type": "string",
            "maxLength": 20
          },
          "ignored": {
            "type": "boolean"
          }
        }
      },
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense",
              "transfer"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "amount": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "description": {
            "type": "string",
            "maxLength": 200,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "note": {
            "type": "string",
            "maxLength": 1000,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "categoryId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "accountId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "toAccountId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z||([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        },
        "required": [
          "type",
          "amount",
          "accountId",
          "date"
        ]
      },
      "TransactionDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "userId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense",
              "transfer"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "amount": {
            "type": "number",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "description": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "note": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "categoryId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "accountId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "toAccountId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "date": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "deletedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true
          },
          "syncVersion": {
            "type": "number",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "createdAt": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "updatedAt": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        },
        "required": [
          "_id",
          "userId",
          "type",
          "amount",
          "accountId",
          "date",
          "deletedAt",
          "syncVersion",
          "createdAt",
          "updatedAt"
        ]
      },
      "TransactionListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string"
                },
                "userId": {
                  "type": "string"
                },
                "type": {
                  "type": "string",
                  "enum": [
                    "income",
                    "expense",
                    "transfer"
                  ]
                },
                "amount": {
                  "type": "number"
                },
                "description": {
                  "type": "string"
                },
                "note": {
                  "type": "string"
                },
                "categoryId": {
                  "type": "string"
                },
                "accountId": {
                  "type": "string"
                },
                "toAccountId": {
                  "type": "string"
                },
                "date": {
                  "type": "string"
                },
                "deletedAt": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "syncVersion": {
                  "type": "number"
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                }
              },
              "required": [
                "_id",
                "userId",
                "type",
                "amount",
                "accountId",
                "date",
                "deletedAt",
                "syncVersion",
                "createdAt",
                "updatedAt"
              ]
            },
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense",
              "transfer"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "amount": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "description": {
            "type": "string",
            "maxLength": 200,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "note": {
            "type": "string",
            "maxLength": 1000,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "categoryId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "accountId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "toAccountId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z||([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        }
      },
...
    }
  }
}